+++ opem/__init__.py
@@ -11,5 +11,6 @@ import opem.Dynamic.Padulles1
 import opem.Dynamic.Padulles2
 import opem.Dynamic.Padulles_Hauer
 import opem.Dynamic.Padulles_Amphlett
+import opem.Dynamic.Chakraborty
 
 __version__ = str(opem.Params.Version)