+++ CHANGELOG.md
@@ -5,6 +5,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- GitHub actions
+- Chakraborty dynamic model
+### Changed
+- Test system modified
+- Padulles-Amphlett dynamic model modified
+- HTML report modified
+- Dockerfile updated
+- `README.md` modified
+- Document modified
 ## [1.2] - 2020-03-12
 ### Added
 - `__version__` variable